Add feature to start a new game
Expand the endGame function to create a newGame button and add a click
event listener to it, to start a new game.
Add a startNewGame function to remove styles that were set during the
game and reset the game data.
Include a div tag to create an overlay in HTML when game ends, and
handle it in the gameEnd and startNewGame functions.
Rename gameOptions to userOptions since it only refers to the user and
is more precise.
In style.css, add styles to the .new-game and .overlay classes.
In index.html, add the class user-option to the user's buttons,
replacing the simple button selector with user-option class in the
userOptions variable.